refactor(datastorage): split dlq/client.go into consumer/drain (Wave 3)
Split pkg/datastorage/dlq/client.go (973 lines) into client.go (trimmed:
Client struct, producer methods, health check), client_consumer.go (Redis
consumer-group read/claim/ack), and client_drain.go (graceful-shutdown DLQ
drain).

Pure code motion, no behavior change, no exported-API change. See audit
report GO-ANTIPATTERN-AUDIT-2026-07-01 §7k-3.
